Conflicts:
	Zend/zend.h
	Zend/zend_API.c
	Zend/zend_API.h
	Zend/zend_compile.c
	Zend/zend_execute.c
	Zend/zend_language_parser.y
	Zend/zend_language_scanner.c
	Zend/zend_multiply.h
	Zend/zend_operators.c
	Zend/zend_operators.h
	Zend/zend_types.h
	ext/standard/array.c
	ext/standard/basic_functions.c
	ext/standard/math.c

+++ ext/standard/php_uuencode.h
@@ -24,8 +24,8 @@
 PHP_FUNCTION(convert_uudecode);
 PHP_FUNCTION(convert_uuencode);
 
-PHPAPI zend_string *php_uudecode(char *src, int src_len);
-PHPAPI zend_string *php_uuencode(char *src, int src_len);
+PHPAPI zend_string *php_uudecode(char *src, size_t src_len);
+PHPAPI zend_string *php_uuencode(char *src, size_t src_len);
 
 #endif /* PHP_UUENCODE_H */
